Switching to east const
This is compared to const west, which is not very consistent, as const always
acts on what is on its left, and if there isn't anything on the left, it will
act on what is on its right. By always placing const on the right of what it
should act on, the rule becomes more consistent.